Added changes to app
